+++ packages/providers/ovhcloud/src/utils.ts
@@ -1,0 +1,2 @@
+export const ENV_VARIABLE_NAME = "OVHCLOUD_API_KEY";
+export const BASE_URL = "https://oai.endpoints.kepler.ai.cloud.ovh.net/v1";