+++ packages/providers/xai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @llamaindex/xai
 
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies [76709c2]
+  - @llamaindex/openai@0.4.22
+
 ## 0.0.24
 
 ### Patch Changes